requirements.txt
fastapi==0.104.1
pydantic==2.5.0
python-dotenv==1.0.0
sqlalchemy==2.0.23